feat: redis 연결 상실 에러 대응 추가

+++ src/common/redis/redis.service.ts
@@ -1,9 +1,11 @@
+/* monte-log-server/src/common/redis/redis.service.ts */
 import { Injectable } from '@nestjs/common';
 import { createClient } from 'redis';
 
 @Injectable()
 export class RedisService {
   private client;
+  private isConnected: boolean;
 
   constructor() {
     this.client = createClient({
@@ -13,21 +15,48 @@ export class RedisService {
 
     this.client
       .connect()
-      .catch((err) => console.error('Redis 연결 오류:', err));
+      .then(() => {
+        this.isConnected = true;
+        console.log('Redis 연결 성공');
+      })
+      .catch((err) => {
+        this.isConnected = false;
+        console.error('Redis 연결 오류:', err);
+      });
   }
 
-  // 세션을 Redis에 저장
   async set(key: string, value: string, ttl: number): Promise<void> {
-    await this.client.set(key, value, { EX: ttl }); // TTL (초 단위)
+    try {
+      if (!this.isConnected) {
+        throw new Error('Redis가 연결되지 않았습니다.');
+      }
+      await this.client.set(key, value, { EX: ttl });
+    } catch (err) {
+      console.error(`Redis set 오류: ${err.message}`);
+    }
   }
 
-  // 세션을 Redis에서 가져오기
   async get(key: string): Promise<string | null> {
-    return await this.client.get(key);
+    try {
+      if (!this.isConnected) {
+        throw new Error('Redis가 연결되지 않았습니다.');
+      }
+      return await this.client.get(key);
+    } catch (err) {
+      console.error(`Redis get 오류: ${err.message}`);
+      return null; // 캐시가 없다고 가정하고 null 반환
+    }
   }
 
-  // 세션을 Redis에서 삭제하기
   async del(key: string): Promise<void> {
-    await this.client.del(key);
+    try {
+      if (!this.isConnected) {
+        throw new Error('Redis가 연결되지 않았습니다.');
+      }
+      await this.client.del(key);
+    } catch (err) {
+      console.error(`Redis del 오류: ${err.message}`);
+      // Redis가 실패해도 계속 진행
+    }
   }
 }
